+++ CHANGELOG.md
@@ -14,7 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - `PlumeTimer` is simplified and now only uses `measure` function.
-- Disabled cache2k from collecting its own statistics
+- Disabled cache2k from collecting its own statistics.
+- Early stopping enabled when no classes needed to update is detected.
+- Feedback regarding files to updated now moved from `INFO` to `DEBUG` logging.
+- Marked `DeltaGraph.apply` as deprecated.
+- `DeltaGraph::toOverflowDb` now only writes to an existing OverflowDB instance.
 
 ## [0.3.9] - 2021-03-23